% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_v2.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot a map with all accidents in a given State in a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer representing a State code number. It is coherced into an integer.}

\item{year}{an integer representing a year.}
}
\description{
This function takes a State number (\code{state.num}) indicator and a year (\code{year}),
and plots a map with all the accident recorded in the given State/year.
}
\details{
If the State number does not exhist in the data for the given year,
returns a message with "invalid STATE number".

If there are no accidents in the given State for the given year, it returns a message
saying "no accidents to plot"
}
\examples{
\dontrun{
fars_map_state(23, 2013)
}

}
