% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_v2.R
\name{fars_read}
\alias{fars_read}
\title{Read .csv data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name or path of the csv file to be read. Accepts compressed csv file.}
}
\value{
This function returns a dataframe with data read from a .csv file.
}
\description{
This function check for the existance of a given \code{filename}, and if the .csv file exist,
it read it and convert it into dataframe.
}
\details{
If the \code{filename} you provide does not existe, the function stops and return an error message
}
\note{
read_csv is evaluated in a context that ignores all ‘simple’ diagnostic messages. See: \code{\link{message}}
}
\examples{
\dontrun{
fars_read("filename.csv")
fars_read("pathname/filename.csv")
}

}
