% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_v2.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple .csv data file for a series a provided years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of integers representing a series of years}
}
\value{
This function returns a dataframe with 2 columns for each year, representing
month number and year of each recorded data
}
\description{
This function take as input a vector of integers (\code{years}) representig years, make the
defalut filename for each year, and read the data in and convert it into data frame.
}
\details{
For each year with no data file, the function returns an error message.
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years(2013:2015)
}

}
