% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{-----------------------------------------------------------------------------------------------
Summarize records per month for a series a provided years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of integers representing a series of years}
}
\description{
This function takes a series of years (\code{years}) as input and returns a tibble with
the number of records per month, by year.
}
\examples{

}
