% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_v2.R
\name{make_filename}
\alias{make_filename}
\title{Print filename for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer. The function coherce the passed value to an integer}
}
\value{
This function returns a character vector containing a formatted combination of text and year number.
}
\description{
This function print a character vector with the filename combining the default
file name and the \code{year} passed as argument to the function
}
\examples{
\dontrun{
make_filename(2017)
make_filename(4034/2)
}

}
